Add EXTRA_CONFIGURE_FLAGS parameter to debian/rules
We are going to pass `--without-zstd` to the gpdb configure flags to
allow GPDB 6.X to build on Ubuntu.

zstd v1.3.1 is provided in Ubuntu 16.04, but is not a recent enough
version to be able to link with GPDB. This forces us to either vendor a
newer version of zstd, or to build without it. We will eventually be
vendoring zstd for the enterprise build of GPDB, but it will likely
never be included with the OSS build on 16.04.

Ubuntu 18.04 provides zstd v1.3.3, which and will at least pass
the gpcontrib/zstd installcheck, so it is possible that we can provide
an OSS build that contains zstd support.

In passing, also rename the ${VERSION} variable to ${GPDB_VERSION} to
disambiguate what version it is referring to.

Co-authored-by: Taylor Vesely <tvesely@pivotal.io>
Co-authored-by: Ben Christel <bchristel@pivotal.io>